Rename folder components/Card to components/TripCard
