+++ components/TripCard/DetailsCard/index.js
